requirements.txt
aiogram==2.22.1
python-dotenv==0.20.0
psycopg2==2.9.6
vk_api==11.9.9
datetime==4.5
asyncio==3.4.3
urllib3==1.26.12
peewee==3.16.2